refactor: split OpenAI overload annotation
